+++ src/core/configure.js
@@ -15,8 +15,12 @@ const ccGlobalStateKeys = ccContext.globalStateKeys;
  * @param {Object} [option.reducer]  you can define multi reducer for a module by specify a reducer
  * @param {Object} [option.moduleReducer]  if you specify moduleReducer and reducer at the same time, the reducer will be ignored!
  * cc will give state module name as moduleReducer key
+ * @param {Object} [option.init]
+ * @param {Object} [option.globalState]  this globalState will been merged to $$global module state
+ * @param {Object} [option.sharedToGlobalMapping]
+ * @param {Object} [option.middlewares]
  */
-export default function (module, state, { singleClass, moduleReducer, reducer, init, globalState, sharedToGlobalMapping } = {}) {
+export default function (module, state, { singleClass, moduleReducer, reducer, init, globalState, sharedToGlobalMapping, middlewares=[] } = {}) {
   if (!ccContext.isCcAlreadyStartup) {
     throw new Error('cc is not startup yet, you can not call cc.configure!');
   }
@@ -99,4 +103,9 @@ export default function (module, state, { singleClass, moduleReducer, reducer, i
     init(helper.getStateHandlerForInit(module));
   }
 
+
+  if (middlewares.length > 0) {
+    const ccMiddlewares = ccContext.middlewares;
+    middlewares.forEach(m => ccMiddlewares.push(m));
+  }
 }